# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      summary: List all students
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: feeStatus
          in: query
          schema:
            type: string
            enum: [paid, unpaid, all]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentListResponse"
    post:
      operationId: createStudent
      tags: [students]
      summary: Add a new student
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentInput"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      summary: Get student by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    put:
      operationId: updateStudent
      tags: [students]
      summary: Update student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StudentUpdate"
      responses:
        "200":
          description: Updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      summary: Delete student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /students/{id}/photo:
    post:
      operationId: uploadStudentPhoto
      tags: [students]
      summary: Upload student photo (base64)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoUploadInput"
      responses:
        "200":
          description: Photo uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}/payments:
    get:
      operationId: getStudentPayments
      tags: [students]
      summary: Get payment history for a student
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Student:
      type: object
      required: [id, name, phone, fatherName, seatNumber, joiningDate, monthlyFee, feeDueDate, feeStatus]
      properties:
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        fatherName:
          type: string
        photoUrl:
          type: ["string", "null"]
        seatNumber:
          type: integer
        joiningDate:
          type: string
        monthlyFee:
          type: number
        feeDueDate:
          type: integer
        feeStatus:
          type: string
          enum: [paid, unpaid, overdue]
        nextDueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isActive:
          type: boolean
        whatsappNumber:
          type: ["string", "null"]
        createdAt:
          type: string
    StudentInput:
      type: object
      required: [name, phone, fatherName, seatNumber, joiningDate, monthlyFee, feeDueDate]
      properties:
        name:
          type: string
        phone:
          type: string
        fatherName:
          type: string
        seatNumber:
          type: integer
        joiningDate:
          type: string
        monthlyFee:
          type: number
        feeDueDate:
          type: integer
        feeStatus:
          type: string
          enum: [paid, unpaid, overdue]
        notes:
          type: string
        whatsappNumber:
          type: string
    StudentUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        fatherName:
          type: string
        seatNumber:
          type: integer
        joiningDate:
          type: string
        monthlyFee:
          type: number
        feeDueDate:
          type: integer
        feeStatus:
          type: string
          enum: [paid, unpaid, overdue]
        notes:
          type: string
        whatsappNumber:
          type: string
        isActive:
          type: boolean
    PhotoUploadInput:
      type: object
      required: [photoBase64]
      properties:
        photoBase64:
          type: string
    StudentListResponse:
      type: object
      required: [students, total, page, limit]
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Payment:
      type: object
      required: [id, studentId, amount, paymentDate, receiptNumber, month, year]
      properties:
        id:
          type: string
        studentId:
          type: string
        studentName:
          type: string
        seatNumber:
          type: integer
        amount:
          type: number
        paymentDate:
          type: string
        receiptNumber:
          type: string
        month:
          type: integer
        year:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...